Moved logconf into app.helpers so that it's actually a part of the application and can be loaded correctly
